Issue #13 Added code to Menu.cs
The added code handles adding and removing (Pushing and Popping) items from the Pseudo-stack and adds the OnPush and OnPop functions that will run at the beginning of the Push and Pop functions and allow Parent classes to overwrite them with their own definitions.

Tested By: Currently Untested.

Reviewer @bpeake13